requirements.txt
openai
pyyaml
pydantic
